Added Enhancements for Active Directory (#1)
Updated the script to use the Active Directory PowerShell module, if available and using Windows Server, to query against a Domain Controller.

Removed the LIST_OF_GROUPS parameter and merged the functionality into GROUP_SEARCH_STRING. So you can now provide a comma separated list of groups and the use of wildcards is allowed.

Updated the readme with all pertinent information. Be sure to read it through to the end.